src/requirements.txt
auto_mix_prep==0.2.0
efficientnet_pytorch==0.7.1
numpy==1.24.2
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==9.4.0
Pillow==10.3.0
pyrootutils==1.0.4
pytorch_lightning==2.1.2
scikit_learn==1.3.2
torch==2.1.1+cu121
torchmetrics==1.2.0
torchvision==0.16.1
